requirements.txt
# ===========================================================
# requirements.txt
# -----------------------------------------------------------
# Lists all dependencies for BusTrack (SBT)
# Used by pip to install consistent versions in venv
# ===========================================================

# --- Core web framework ---
fastapi==0.115.0                 # Main FastAPI library (API framework)
uvicorn[standard]==0.31.1        # ASGI server with auto-reload and CLI

# --- Database and ORM ---
SQLAlchemy==2.0.36               # ORM layer for defining models and queries
alembic==1.13.3                  # Database migrations for SQLAlchemy models

# --- Data validation ---
pydantic==2.9.2                  # Model validation (FastAPI uses it for schemas)
pydantic-settings==2.6.1         # Manage environment variables and config files

# --- Templates and multipart handling ---
jinja2==3.1.4                    # Template engine for HTML dashboards
python-multipart==0.0.9          # Form and file uploads support in FastAPI

# --- Utility packages ---
python-dotenv==1.0.1             # Loads .env values into environment
python-dateutil==2.9.0.post0     # Easier datetime parsing for reports

# --- Optional features / helpers ---
passlib[bcrypt]==1.7.4           # Password hashing (for future auth)
email-validator==2.2.0           # Validate email addresses
requests==2.32.3                 # HTTP client (if you need external API calls)

# --- Developer / testing tools ---
pytest==8.3.2                    # Testing framework
httpx==0.27.2                    # Async HTTP client for FastAPI testing
black==24.10.0                   # Code formatter (keeps style consistent)

# --- Production (add later when deploying) ---
gunicorn==23.0.0                 # Production WSGI/ASGI process manager
psycopg2-binary        # PostgreSQL database driver
